+++ packages/strapi-admin/admin/src/components/Roles/ConditionsSelect/CustomDropdownIndicator/index.js
